third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/transferToGPUTexture-usage-flags.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-promise-test.js"></script>
<script src="./webgpu-helpers.js"></script>

<h1>Canvas2D WebGPU Transfer: transferToGPUTexture-usage-flags</h1>
<p>transferToGPUTexture() creates a texture honoring the requested usage flags.
</p>

<script>

runCanvasTestsInWorker({dependencies: ['./webgpu-helpers.js']});

/**
 * transferToGPUTexture() should create a texture which honors the requested
 * usage flags.
 */
for (const {name, usageFlag, usageCode} of [
    {name: "COPY_SRC",
     usageFlag: GPUTextureUsage.COPY_SRC,
     usageCode: (device, tex) => copyOnePixelFromTextureAndSubmit(device, tex)},
    {name: "COPY_DST",
     usageFlag: GPUTextureUsage.COPY_DST,
     usageCode: (device, tex) => copyOnePixelToTextureAndSubmit(device, tex)},
    {name: "TEXTURE_BINDING",
     usageFlag: GPUTextureUsage.TEXTURE_BINDING,
     usageCode: (device, tex) => createBindGroupUsingTexture(device, tex)},
    {name: "RENDER_ATTACHMENT",
     usageFlag: GPUTextureUsage.RENDER_ATTACHMENT,
     usageCode: (device, tex) => clearTextureToColor(
          device, tex, { r: 1.0, g: 1.0, b: 1.0, a: 1.0 })},
]) {
  /**
   * Usage produce an error if flag isn't specified in transferToGPUTexture().
   */
  canvasPromiseTest(async (canvas) => {
    const adapter = await getWebGPUAdapter();
    // Skip this test on Mac Swiftshader due to "Invalid Texture" errors.
    if (isMacSwiftShader(adapter.info)) {
      return;
    }

    const device = await getWebGPUDevice(adapter);
    const ctx = canvas.getContext('2d');

    // Transfer texture without any usage flags.
    const tex = ctx.transferToGPUTexture({device: device, usage: 0});
    assert_equals(tex.usage, 0);

    // Trying to perform the usage should fail with a GPUValidationError.
    device.pushErrorScope('validation');
    usageCode(device, tex);
    assert_not_equals(await device.popErrorScope(), null);
  }, name + ' won\'t work if usage isn\'t specified in transferToGPUTexture()');

  /**
   * Usage succeed if corresponding flag is specified in transferToGPUTexture().
   */
  canvasPromiseTest(async (canvas) => {
    const adapter = await getWebGPUAdapter();
    // Skip this test on Mac Swiftshader due to "Invalid Texture" errors.
    if (isMacSwiftShader(adapter.info)) {
      return;
    }

    const device = await getWebGPUDevice(adapter);
    const ctx = canvas.getContext('2d');

    // Transfer texture with the usage flag under test.
    const tex = ctx.transferToGPUTexture({device: device, usage: usageFlag});
    assert_equals(tex.usage, usageFlag);

    // Performing the usage should succeed.
    device.pushErrorScope('validation');
    usageCode(device, tex);
    assert_equals(await device.popErrorScope(), null);
  }, name + ' will work if usage is specified in transferToGPUTexture()');
}

</script>
